2D Fighting Game - MB2
Game made using game maker. Player one uses WASD to move and q for
fireball and e for kick (there's a typo in the tutorial stage). Player
two uses control for fireball and shift for attack. You can only attack
when the timer is the same colour as you.  While in the air you can hit
down to fall fast. You can hit down and a direction to dash while in the
air. Your objective is to boot your enemy past the last stage.